fix: add test for child count updater deadlock
